Chapter 13.4. Mocking Dependencies
We’ve updated the application struct so that instead of the snippets and users fields having the concrete types *mysql.SnippetModel and *mysql.UserModel they are interfaces instead.

So long as an object has the necessary methods to satisfy the interface,
we can use them in our application struct. Both our ‘real’ database
models (under the pkg/models/mysql package) and mock database models
(under the pkg/models/mock package) satisfy the interfaces, so we can
now use them interchangeably.